Lab 8

+++ database/blueprints/blpr_note.py
@@ -1,15 +1,12 @@
-from sqlalchemy import func
-
-from database.flask_ini import app
-
-from database.models import note, note_log, action
-
+from database.models import note, note_log, user, action, connected_user
 from database.db_utils import *
-
 from database.schemas import NoteSchema
+
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 
 @app.route("/note", methods=["POST"])
+@jwt_required()
 @db_lifecycle
 @session_lifecycle
 def create_note():
@@ -17,8 +14,17 @@ def create_note():
     obj = note(**data)
     act = action(name="created note")
 
-    if request.json.get("user_id", None) is None:
+    user_id = request.json.get("user_id", None)
+
+
+    if user_id is None:
         raise InvalidUsage("user_id is not present", status_code=400)
+
+    user_obj = user.query.filter_by(id=user_id).first()
+    current_user_email = get_jwt_identity()
+
+    if current_user_email != user_obj.email:
+        return jsonify("Access denied", 402)
 
     session.add(act)
     session.add(obj)
@@ -26,10 +32,28 @@ def create_note():
 
     log = note_log(obj.id, request.json.get("user_id", None), act.id)
 
+    cu = connected_user(user_id=user_id, note_id=obj.id)
+    session.add(cu)
     session.add(log)
 
     return jsonify(NoteSchema().dump(obj))
 
+
+@app.route('/note/add/<int:input_user_id>,<int:input_note_id>', methods=["PUT"])
+@jwt_required()
+@db_lifecycle
+@session_lifecycle
+def add_user(input_user_id, input_note_id):
+    current_user_email = get_jwt_identity()
+    user_obj = user.query.filter_by(email=current_user_email).first()
+    note_obj = note.query.filter_by(id=input_note_id).first()
+
+    if user_obj.id != note_obj.user_id:
+        return jsonify("Access denied", 402)
+
+    cu = connected_user(user_id=input_user_id, note_id=input_note_id)
+    session.add(cu)
+    return jsonify("User added")
 
 @app.route('/note', methods=["GET"])
 def get_notes():
@@ -37,8 +61,17 @@ def get_notes():
 
 
 @app.route("/note/<int:Id>", methods=["GET"])
+@jwt_required()
+@db_lifecycle
 def get_note_by_Id(Id):
-    return get_obj_by_Id(NoteSchema, note, Id)
+    current_user_email = get_jwt_identity()
+    user_obj = user.query.filter_by(email=current_user_email).first()
+
+    for cu in connected_user.query.all():
+        if cu.note_id == Id and user_obj.id == cu.user_id:
+            return get_obj_by_Id(NoteSchema, note, Id)
+
+    return jsonify("Access denied", 402)
 
 
 @app.route("/note/<string:tag>", methods=["GET"])
@@ -50,11 +83,21 @@ def get_notes_by_tag(tag):
 @app.route("/note/<int:Id>", methods=["PUT"])
 @db_lifecycle
 @session_lifecycle
+@jwt_required()
 def upd_note_by_Id(Id):
-    new_data = NoteSchema().load(request.get_json())
+    current_user_email = get_jwt_identity()
+    user_obj = user.query.filter_by(email=current_user_email).first()
 
-    if request.json.get("user_id", None) is None:
-        raise InvalidUsage("user_id not present", status_code=404)
+    check = False
+
+    for cu in connected_user.query.all():
+        if cu.note_id == Id and user_obj.id == cu.user_id:
+            check = True
+
+    if not check:
+        return jsonify("Access denied", 402)
+
+    new_data = NoteSchema().load(request.get_json())
 
     if session.query(note_log).filter_by(user_id=request.json.get("user_id", None)).first() is None:
         if session.query(note_log).filter_by(note_id=Id).distinct("user_id").count() > 4:
@@ -81,5 +124,20 @@ def upd_note_by_Id(Id):
 
 
 @app.route("/note/<int:Id>", methods=["DELETE"])
+@db_lifecycle
+@session_lifecycle
+@jwt_required()
 def delete_note_by_id(Id):
+    current_user_email = get_jwt_identity()
+    user_obj = user.query.filter_by(email=current_user_email).first()
+    note_obj = note.query.filter_by(id=Id).first()
+
+    if note_obj.user_id != user_obj.id:
+        return jsonify("Access denied", 402)
+
+    session2 = Session()
+    for cu in connected_user.query.all():
+        if cu.note_id == Id:
+            session2.delete(cu)
+    session2.commit()
     return delete_obj_by_id(NoteSchema, note, Id)
