+++ database/blueprints/blpr_user.py
@@ -1,18 +1,13 @@
+import bcrypt
 import datetime
 
-import bcrypt
-
 from database.models import user
-
 from database.db_utils import *
-
 from database.schemas import UserSchema
 
-from flask_jwt_extended import create_access_token
+from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity
 
-from flask_jwt_extended import jwt_required
-
-from flask_jwt_extended import get_jwt_identity
+app.config['SECRET_KEY'] = 'secret'
 
 
 @app.route("/user", methods=["POST"])
@@ -32,29 +27,21 @@ def create_user():
     return jsonify(UserSchema().dump(obj))
 
 
-@app.route("/user/login", methods=["POST"])
+@app.route("/login", methods=["GET"])
 @db_lifecycle
 def login():
-    email = request.json.get('email', None)
-    password = request.json.get('password', None)
+    auth = request.authorization
 
-    user_obj = user.query.filter_by(email=email).first()
+    if not auth or not auth.username or not auth.password:
+        return jsonify("Couldn't verify", 401)
 
-    if bcrypt.checkpw(password.encode("utf-8"), user_obj.password.encode("utf-8")):
+    user_obj = user.query.filter_by(email=auth.username).first()
 
-        access_token = create_access_token(identity=str(user.id), expires_delta=datetime.timedelta(days=7))
+    if user_obj.password == auth.password:
+        access_token = create_access_token(identity=auth.username, expires_delta=datetime.timedelta(days=7))
+        return jsonify({'token': access_token})
 
-        return jsonify({'token': access_token}), 200
-        # return jsonify(UserSchema().dump(user_obj))
-    else:
-        raise InvalidUsage("Bad username or password", status_code=401)
-
-
-@app.route("/protected", methods=["GET"])
-@jwt_required()
-def protected():
-    current_user = get_jwt_identity()
-    return jsonify(logged_in_as=current_user), 200
+    raise InvalidUsage("Unexisting username or password", status_code=401)
 
 
 @app.route('/user', methods=["GET"])
@@ -63,21 +50,52 @@ def get_users():
 
 
 @app.route("/user/<int:Id>", methods=["GET"])
+@jwt_required()
+@db_lifecycle
 def get_user_by_Id(Id):
+    current_user_email = get_jwt_identity()
+    user_obj = user.query.filter_by(id=Id).first()
+
+    if current_user_email != user_obj.email:
+        return jsonify("Access denied", 402)
+
     return get_obj_by_Id(UserSchema, user, Id)
 
 
 @app.route("/user/<string:username>", methods=["GET"])
+@jwt_required()
+@db_lifecycle
 def get_user_by_name(username):
-    obj = user.query.filter_by(username=username).first()
-    return jsonify(UserSchema().dump(obj))
+    current_user_email = get_jwt_identity()
+    user_obj = user.query.filter_by(username=username).first()
+
+    if current_user_email != user_obj.email:
+        return jsonify("Access denied", 402)
+
+    return jsonify(UserSchema().dump(user_obj))
 
 
 @app.route("/user/<int:Id>", methods=["PUT"])
+@jwt_required()
+@db_lifecycle
 def upd_user_by_Id(Id):
+    current_user_email = get_jwt_identity()
+    user_obj = user.query.filter_by(id=Id).first()
+
+    if current_user_email != user_obj.email:
+        return jsonify("Access denied", 402)
+
     return upd_obj_by_Id(UserSchema, user, Id)
 
 
 @app.route("/user/<int:Id>", methods=["DELETE"])  # delete user by id
+@jwt_required()
+@db_lifecycle
 def delete_user_by_id(Id):
+    current_user_email = get_jwt_identity()
+    user_obj = user.query.filter_by(id=Id).first()
+
+    if current_user_email != user_obj.email:
+        return jsonify("Access denied", 402)
+
     return delete_obj_by_id(UserSchema, user, Id)
